% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_p.R
\name{boot_p}
\alias{boot_p}
\title{Simulate and do inference for Pr(X > x) at different thresholds}
\usage{
boot_p(data, var, thresh, its = 1000, ci = 0.95)
}
\arguments{
\item{data}{A data object. Can be either grouped or un-grouped.}

\item{var}{The variable for which Pr(X > x) will be computed.}

\item{thresh}{A numerical threshold for which Pr(X > x) is computed. If missing, an error will be returned.}

\item{its}{The number of bootstrap iterations. The default is 1,000.}

\item{ci}{A value between 0 and 1 indicating the quantile confidence interval level to be returned. 0.95 is the default.}
}
\value{
A data frame with the following values:
\describe{
\item{estimate}{The estimated Pr(X > x).}
\item{lower}{The lower bound of the bootstrapped confidence interval.}
\item{upper}{The upper bound of the bootstrapped confidence interval.}
}
}
\description{
This function allows you to estimate Pr(X > x) for some variable, potentially grouped by any number of factors, and
perform bootstrapping for statistical inference.
}
\examples{
library(dplyr)
wars |>
  group_by(post1950) |>
  boot_p(fat, thresh = 1e06, ci = .84)

}
