% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llplot.R
\name{llplot}
\alias{llplot}
\title{Make a log-log plot of the empirical Pr(X > x) given x}
\usage{
llplot(data, x, by, show_fit = FALSE)
}
\arguments{
\item{data}{A data object for the variable you want to plot.}

\item{x}{The variable you want to show Pr(X > x).}

\item{by}{An optional grouping variable. Should be a character or factor.}

\item{show_fit}{If \code{TRUE} the plot will include an inverse Burr fit for the data. Default is \code{FALSE}.}
}
\description{
This function allows you to plot a
log-log plot of Pr(X > x) given x.
}
\examples{
llplot(wars, fat, show_fit = T)
}
