% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_p.R
\name{boot_p}
\alias{boot_p}
\title{Simulate Pr(X > x) for an inverse Burr model outcome}
\usage{
boot_p(data, var, thresh, its = 1000, ci = 0.95)
}
\arguments{
\item{data}{A data object. Can be either grouped or un-grouped.}

\item{var}{The variable for which Pr(X > x) will be computed.}

\item{thresh}{A numerical threshold for which Pr(X > x) is computed. If missing, an error will be returned.}

\item{its}{The number bootstrap iterations. The default is 1,000.}

\item{ci}{A value between 0 and 1 indicating the quantile confidence interval level to be returned. 0.95 is the default.}
}
\description{
This function allows you to simulate random draws from an
inverse Burr model fit using the \code{ibm()} function.
}
