% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibm_sim.R
\name{ibm_sim}
\alias{ibm_sim}
\title{Simulate draws from a fitted inverse Burr model}
\usage{
ibm_sim(model, its = 1000, newdata = NULL)
}
\arguments{
\item{model}{A model fit with the \code{ibm()} function.}

\item{its}{The number of random draws. The default is 1,000.}

\item{newdata}{An optional data frame with new covariate values to use for simulated values.}
}
\description{
This function allows you to simulate random draws from an
inverse Burr model fit using the \code{ibm()} function.
}
